chore: reroute types folder in order to keep it consistent with other exports
