+++ src/types/UKrew.ts
@@ -1,0 +1,17 @@
+export interface UKrewStudent {
+  firstName: string;
+  middleName?: string;
+  lastName: string;
+  email: string;
+  department: string;
+  role: string;
+  major: string;
+  phoneNumber: string;
+  graduationSemester: string;
+  graduationYear: string;
+  session: string;
+  linkedIn?: string;
+  photoUpload?: string;
+  image?: string;
+  portfolioLink?: string;
+}
